[E3DB-1133] Update to using and being compatible by the go module tool. - Shuffle examples around so they are present in GoDoc - Add some doc to the quixotic Q type. - Ensure all exported fields and methods have a comment and are formatted correctly.
