--- glide.lock
@@ -1,41 +1,0 @@
-hash: c2d1acf6bcfa6d96138ef3c3f6a449b8f0a4aec941bec1625a35571fc6a75fbe
-updated: 2017-04-27T13:22:49.3885103-07:00
-imports:
-- name: github.com/golang/protobuf
-  version: 2bba0603135d7d7f5cb73b2125beeda19c09f4ef
-  subpackages:
-  - proto
-- name: github.com/gorilla/websocket
-  version: 3ab3a8b8831546bd18fd182c20687ca853b2bb13
-- name: github.com/jawher/mow.cli
-  version: d3ffbc2f98b83e09dc8efd55ecec75eb5fd656ec
-- name: github.com/mitchellh/go-homedir
-  version: b8bc1bf767474819792c23f32d8286a45736f1c6
-- name: golang.org/x/crypto
-  version: c7af5bf2638a1164f2eb5467c39c6cffbd13a02e
-  subpackages:
-  - curve25519
-  - nacl/box
-  - nacl/secretbox
-  - poly1305
-  - salsa20/salsa
-- name: golang.org/x/net
-  version: da118f7b8e5954f39d0d2130ab35d4bf0e3cb344
-  subpackages:
-  - context
-- name: golang.org/x/oauth2
-  version: a6bd8cefa1811bd24b86f8902872e4e8225f74c4
-  subpackages:
-  - clientcredentials
-  - internal
-- name: google.golang.org/appengine
-  version: 170382fa85b10b94728989dfcf6cc818b335c952
-  subpackages:
-  - internal
-  - internal/base
-  - internal/datastore
-  - internal/log
-  - internal/remote_api
-  - internal/urlfetch
-  - urlfetch
-testImports: []
