--- glide.yaml
@@ -1,13 +1,0 @@
-package: github.com/tozny/e3db-go
-import:
-- package: github.com/jawher/mow.cli
-- package: github.com/mitchellh/go-homedir
-- package: golang.org/x/crypto
-  subpackages:
-  - nacl/box
-  - nacl/secretbox
-- package: golang.org/x/oauth2
-  subpackages:
-  - clientcredentials
-- package: github.com/gorilla/websocket
-  version: ~1.1.0
