+++ go.mod
@@ -1,0 +1,10 @@
+module github.com/tozny/e3db-go/v2
+
+require (
+	github.com/gorilla/websocket v1.4.0
+	github.com/jawher/mow.cli v1.0.4
+	github.com/mitchellh/go-homedir v1.0.0
+	golang.org/x/crypto v0.0.0-20181112202954-3d3f9f413869
+	golang.org/x/net v0.0.0-20181114220301-adae6a3d119a // indirect
+	golang.org/x/oauth2 v0.0.0-20181120190819-8f65e3013eba
+)
